requirements.txt
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
httpx>=0.24.0
pydantic>=2.0.0
pytest>=7.0.0
pyyaml>=6.0
jupyter>=1.0.0
